requirements.txt
moviepy==1.0.0
numpy==1.11.2
scipy==0.18.1
pillow==3.4.2
PyVimeo==1.0.11